Icon in background
The QuanPic Icon is now added as the background. Unfortunately some
corrections still need to be made as other objects are in the way so
that the logo cannot fully be seen. Also, the QuanPic logo will be
adjusted to each phone's resolution so it might look a bit weird.